+++ src/lib.rs
@@ -12,5 +12,6 @@ pub mod time;
 pub use crate::enumerate::enumerate;
 pub use crate::input::input;
 pub use crate::iterable::Iterable;
+pub use crate::reversed::reversed;
 pub use crate::reversed::Reversed;
 pub use crate::sorted::sorted;