+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add function: `rspy::sorted`
 - Add trait: `rspy::Reversed`
+- Add function: `rspy::reversed`
 
 ## v0.1.10